+++ tests/Flare.Components.Tests/Menu/FlareMenuTests.cs
@@ -5,6 +5,12 @@ namespace Flare.Components.Tests;
 
 public class FlareMenuTests : FlareTestContext
 {
+    // The activator slot is a template: the menu hands it the ARIA a menu button has to carry. These
+    // tests only need something clickable in the wrapper, so they ignore the context they are given -
+    // FlareMenuActivatorAriaTests is where the attributes themselves are checked.
+    private static RenderFragment<FlareMenuActivatorContext> MenuActivator(string html) =>
+        _ => b => b.AddMarkupContent(0, html);
+
     [Fact]
     public void RendersRootDiv()
     {
@@ -33,8 +39,7 @@ public class FlareMenuTests : FlareTestContext
     public void ClickActivator_OpensMenuPanel()
     {
         var cut = Render<FlareMenu>(p => p
-            .Add(x => x.Activator, b =>
-                b.AddMarkupContent(0, "<button id=\"act\">Open</button>")));
+            .Add(x => x.Activator, MenuActivator("<button id=\"act\">Open</button>")));
 
         cut.Find($".{Css.Classes.Menu.Activator}").Click();
 
@@ -45,8 +50,7 @@ public class FlareMenuTests : FlareTestContext
     public void MenuPanel_HasRoleMenu()
     {
         var cut = Render<FlareMenu>(p => p
-            .Add(x => x.Activator, b =>
-                b.AddMarkupContent(0, "<button>Open</button>")));
+            .Add(x => x.Activator, MenuActivator("<button>Open</button>")));
 
         cut.Find($".{Css.Classes.Menu.Activator}").Click();
 
@@ -57,8 +61,7 @@ public class FlareMenuTests : FlareTestContext
     public void RendersMenuItems_WhenOpen()
     {
         var cut = Render<FlareMenu>(p => p
-            .Add(x => x.Activator, b =>
-                b.AddMarkupContent(0, "<button>Open</button>"))
+            .Add(x => x.Activator, MenuActivator("<button>Open</button>"))
             .AddChildContent<FlareMenuItem>(mi =>
                 mi.AddChildContent("Item One")));
 
@@ -71,8 +74,7 @@ public class FlareMenuTests : FlareTestContext
     public void DefaultAnchor_HasBottomLeftClass()
     {
         var cut = Render<FlareMenu>(p => p
-            .Add(x => x.Activator, b =>
-                b.AddMarkupContent(0, "<button>Open</button>")));
+            .Add(x => x.Activator, MenuActivator("<button>Open</button>")));
 
         cut.Find($".{Css.Classes.Menu.Activator}").Click();
 
@@ -83,8 +85,7 @@ public class FlareMenuTests : FlareTestContext
     public void RendersBackdrop_WhenOpen()
     {
         var cut = Render<FlareMenu>(p => p
-            .Add(x => x.Activator, b =>
-                b.AddMarkupContent(0, "<button>Open</button>")));
+            .Add(x => x.Activator, MenuActivator("<button>Open</button>")));
 
         cut.Find($".{Css.Classes.Menu.Activator}").Click();